add wkt.mk to make building WKT libraries easier
This adds a new file, `wkt.mk`, that can be included by client makefiles
to build WKT libraries.

To build a wkt library, e.g., `libfoo-wkt.a`, simply include `wkt.mk`,
define a variable `foo.wkt_src`, and add the WKT library as a make target.
For example:

    include /path/to/wkt.mk
    foo.wkt_src = wkt_shapes.h wkt_colors.h wkt_numbers.h
    all: libfoo-wkt.a

This will generate Fortran modules and .o files from the WKT files, add
them to the WKT library, and create the library.